# Faster queue free