Calling queue_free() for large numbers of siblings could previously be pathologically slow, with the time taken rising exponentially with number of children. This looked partly due to ordered_remove from the child list and notifications.

This PR identifies objects that are nodes, and deletes children flagged for deletion in reverse order. This minimizes the costs of reordering.

Fixes #61929

## Performance testing, number of child nodes deleted

**50000 nodes** before 137017 ms, after 422 ms seconds (**325x** faster)
**20000 nodes** before 19425 ms, after 250 ms seconds (**78x** faster)
**5000 nodes** before 1228 ms, after 169 ms seconds (**7.3x** faster)
**1000 nodes** before 194 ms, after 150 ms seconds (**1.3x** faster)

(at lower node numbers the measurement from the MRP in #61929 becomes less useful as there are some fixed delays, however the pattern is that the benefit really starts to show with a decent number of children)

## Notes
* It might be even nicer to sort the `delete_queue`, but there seems no efficient way of getting the child number from a node.
* The element "id" could be removed from the linked list during the first stage, but it is probably not worth it provided the `ObjectDB::get_instance()` call is efficient for deleted objects.
* This could alternatively be done as part of the lower iteration section of the function (instead of separately), which might save some `ObjectDB::get_instance()` calls, but might be more tricky to turn off efficiently.
* The flip side of this PR is that it can add a small cost when e.g. a single child is being deleted from among 1000 siblings - the siblings will be traversed *just in case* they are queued for deletion. This will usually be avoided however by the check for the overall queue size.
* The 2 magic numbers are to some extent guessed. I've set them rather conservatively so in most cases during gameplay they won't activate, but during level shutdown / closing the editor etc they will help prevent long pauses.
* Same principle can be applied in 4.x as the code is similar.

## Update
It turns out that after the fix the bottleneck was in the BVH doing collision detection of all the boxes overlapping creating a lot of pairs. When the boxes were hidden, the artificial BVH bottleneck is removed and the new PR is _much_ faster (i.e. 100x faster).

<!--
Pull requests should always be made for the `master` branch first, as that's
where development happens and the source of all future stable release branches.

Relevant fixes are cherry-picked for stable branches as needed.

Do not create a pull request for stable branches unless the change is already
available in the `master` branch and it cannot be easily cherry-picked.
Alternatively, if the change is only relevant for that branch (e.g. rendering
fixes for the 3.2 branch).
-->