@goatchurchprime 2022-06-13
This seems like an overly-narrow fix for an interesting performance issue that's going to occur every time an object is deleted, even if it's not always big enough to be noticed and isolated.  

Also, won't it make things worse if I am already calling queue_free() on objects in reverse order to work around this problem?

There must be something wrong in the implementation if it makes this much difference by simply changing the order of deletion.  Is there a low level function missing its binary search?

I am also skeptical about a batch delete function that simply loops through a single object delete, if a lot of time can be wasted through relinking and shuffling objects that are also due to be deleted.  There is a memdelete_arr() function, but this also does it by looping through its individual records too because it doesn't know enough about what the object types are.