@lawnjelly 2022-06-13
> an interesting performance issue that's going to occur every time an object is deleted, even if it's not always big enough to be noticed and isolated.

It only seems to occur when deleting multiple objects. Deleting a single object (that has many children) doesn't seem to incur the same cost, presumably because the children of this object are deleted in bulk, and don't get reshuffled and sent notifications multiple times etc.

> I am also skeptical about a batch delete function that simply loops through a single object delete, if a lot of time can be wasted through relinking and shuffling objects that are also due to be deleted.

Absolutely, a bulk delete function could be made more efficient, see particularly  https://github.com/godotengine/godot/issues/61929#issue-1268226791 :

> Additionally, if there is a notification on re-ordering, perhaps it can be sent once (after all deletes) instead of after each deletion.

The API doesn't currently offer a "multiple delete" afaik, only `queue_free()` can be called multiple times. And we are constrained by history here. However we can make the machinery more efficient behind this API in a number of different ways.

This PR is currently a super simple way of addressing this and getting an easy "bang for your buck", but if desired it could be made more comprehensive - there are several obvious ways this could be improved further. But this is something we will typically decide at a PR meeting. It is good practice to make a simple solution first before investing significant time, as often reduz will prefer the simple solution, and this saves develop time and frustration.

There is a trade off, with code complexity versus benefit. There's also a difference between theoretical benefit and practical benefit. The PR currently is simple and has a lot of practical benefit. We could increase the complexity and theoretical benefit further but maybe not gain much in terms of practical benefit. The Godot philosophy on the whole errs on the side of simple maintainable code where possible (over outright performance).

> Also, won't it make things worse if I am already calling queue_free() on objects in reverse order to work around this problem?

In theory, but in practice it will make little difference, because the first child it hits in the list to be deleted will delete all the queued children of the parent, then the later `get_instance()` will fail (this relies on `get_instance()` being reasonably fast, but it does seem to be fast in my tests, relative to the costs of deletion).