@lawnjelly 2022-06-27
Some testing of alternative strategies based on discussion with reduz:
### Dirty flag for "moved in parent"
Based on the hypothesis that a lot of the cost might be the `moved in parent` notifications being called multiple times, instead of calling `notification(NOTIFICATION_MOVED_IN_PARENT)` directly during `remove_child()`, we set a flag on the node (if not set already) and add to a list, then flush this list once (per frame / iteration), ensuring that there is only a single NOTIFICATION_MOVED_IN_PARENT per node rather than multiple.

_Timings (5000 nodes, debug)_
This PR (reverse iteration approach) 280 ms
Dirty flag 2526 ms
Existing 3.x 3424 ms

So using a dirty flag seems to be faster than the existing 3.x branch, but not nearly as fast as the reverse iteration approach (for this particular benchmark test anyway, which is worst case).

One advantage of the dirty flag (over reverse iteration alone) is it will also work when calling `remove_child()` directly, rather than via `queue_free()`. However a disadvantage is that the dirty flag needs to be made thread safe, and creates the potential for bugs because the timing for the NOTIFICATION_MOVED_IN_PARENT will have been altered (it will be called in a deferred fashion rather than immediately). This could potentially cause knock on bugs so is less safe in terms of regressions.

The dirty flag could potentially be used _in addition_ to the reverse iteration in this PR. I tested this combined approach, and it is slightly slower than just reverse iteration on its own, but as said above could be helpful in more cases. But this should probably be added in a separate PR, if we were to go with this.